Renamed StyleMap and Style to OptionMap and Opts respectively
To allow plot parameters to be configured via the same mechanism as
styles, StyleMap and Styles have been renamed to OptionMap and Opts. The
new OptionGroup component allows easy definition of multiple categories
of options with a shared set of keys. The 'options' object is the
default instance of this class that is a handle allowing plotting and
style options to be set. The file has been renamed accordingly.