Removed the origin style option from the raster plots
The origin option flips the display at the level of imshow without
flipping the tick labels appropriately. This breaks the link between the
semantics of the data (i.e the sheetcoordinate system) and what is
displayed. This is likely to cause a lot of confusion when you can
either use the invert_yaxis plot option or flip the data
itself. Discussed further in issue #76