Moved merge_options back to StoreOptions from OptionTree
This move has been reconsidered as the use of attribute access on
OptionTree means we want to minimise the number of methods on this
class, even if they would otherwise make sense there.

Utilities for manipulating OptionTrees should be implemented on
StoreOptions (the docstring has been updated to reflect this).