REL+BUG: Fix panel requirement on Python 3.5
The current release of Holoviews still claims to support Python 3.5.
However, it can not be installed on Python 3.5 with its current
installation requirements. The reason for this is an anomaly in
the release history of panel. We work around this problem by
specifying a maximum version requirement of panel for holoviews
when installing on Python 3.5.
More details of the problem are given below.

The new major release of bokeh, v2.0.0, dropped support for Python
3.5. This new python version requirement was specified in its
setup.py requirements.
To support the new bokeh API, panel was changed with what was
effectively a new major release with v0.9.0. This version of
bokeh is only for bokeh>=2.0.0, and this requirement was specified
in its setup.py requirements. However, panel did not change its
Python version requirements to >=3.6 until version 0.9.4.
This means that whenever one tries to install the latest available
version of panel on Python 3.5, pip will attempt to install version
0.9.3, and then fail because it is not possible to support the
bokeh>=2.0.0 requirement on Python 3.5. Consequently, all packages
which require panel and continue to support Python 3.5 need to
include an extra specification for the panel version on Python <3.6.