Fix Labels on categorical data
Fix Labels when working with a categorical axis like:

    hv.Labels([("hello", 12, "text"), ("world", 18, "text2")])

This is especially important as Labels can be used to annotate bar
graphs.

Since matplotlib does not correctly infer the axis units for ax.text(),
we have to help it by adding an invisible scatter plot, so that the axes
are correctly initialized.

Fixes #4992