Fix unhandled numpy.round overflow return value in core/util.py:bound_range(...) (#5095)
* Core/Util: Fix unhandled np.round overflow return

The function numpy.round return for the input
np.float64(2.6558061446181644e+295) the output
numpy.inf. This should lead to
density = full_precision_density.

* Core/Util: Supress numpy.round overflow error

The function numpy.round produces for the input
numpy.float64(2.6558061446181644e+295) an
FloatingPointError with the message
"overflow encountered in multiply". This error
should be suppressed because its an internal
computation by holoview.

* Use context manager

Co-authored-by: christoph.weiss@gtd-gmbh.de <christoph.weiss@gtd-gmbh.de>
Co-authored-by: Philipp Rudiger <prudiger@anaconda.com>