Added unit tests for "regular" Bokeh `QuadMesh` plots.
This unit tests checks that any extra vdims provided when instantiating a `QuadMesh` with 1-D coordinate arrays are correctly passed to the Bokeh `ColumnDataSource` for use in hover labels. This feature is fully supported and functioning properly in HoloViews 1.15.4 (so the test should pass), unlike for 2-D coordinates wherein any extra vdims are lost.