Added missing support for supplying extra vdims when creating Bokeh `QuadMesh` plots from 2-D coordinate arrays.
This change allows for the inclusion of additional variables in the hovertool for non-rectilinear geometries, which has long been supported for simple geometries defined by 1-D coordinate arrays. This fix only required very minor changes to the `QuadMeshPlot.get_data` method, and no other functionality should be affected.

All tests now pass, including the 2 new ones added to test this functionality, as well as the 3rd new test added to ensure that the existing behavior for 1-D coordinates was not affected.