Moved the logic of whether the data needs to be transposed to the `_collect_hover_data` method.
This involved converting the staticmethod to an instance method, changing the `transpose` argument to `irregular`, and then adding a line to derive `transpose` from `invert_axes` and `irregular`.

This change was made to increase clarity by putting the unintuitive step closer to where it is used, and right next to the docstring explaining it.